Remove Amp\Stream\interval and add delay option to fromIterable
