+++ lib/functions.php
@@ -1,7 +1,6 @@
 <?php
 
 namespace Amp {
-
     use React\Promise\PromiseInterface as ReactPromise;
 
     /**
@@ -92,7 +91,6 @@ namespace Amp {
 }
 
 namespace Amp\Promise {
-
     use Amp\Loop;
     use Amp\Deferred;
     use Amp\MultiReasonException;
@@ -491,11 +489,11 @@ namespace Amp\Promise {
 }
 
 namespace Amp\Stream {
-
     use Amp\Coroutine;
     use Amp\Emitter;
     use Amp\Listener;
     use Amp\Loop;
+    use Amp\Pause;
     use Amp\Producer;
     use Amp\Promise;
     use Amp\Stream;
@@ -505,19 +503,24 @@ namespace Amp\Stream {
      * Creates a stream from the given iterable, emitting the each value. The iterable may contain promises. If any promise
      * fails, the stream will fail with the same reason.
      *
-     * @param array|\Traversable $iterable
+     * @param array|\Traversable $iterable Elements to emit.
+     * @param int $delay Delay between element emissions.
      *
      * @return \Amp\Stream
      *
      * @throws \TypeError If the argument is not an array or instance of \Traversable.
      */
-    function fromIterable(/* iterable */ $iterable): Stream {
+    function fromIterable(/* iterable */ $iterable, int $delay = 0): Stream {
         if (!$iterable instanceof \Traversable && !\is_array($iterable)) {
             throw new UnionTypeError(["array", "Traversable"], $iterable);
         }
 
-        return new Producer(function (callable $emit) use ($iterable) {
+        return new Producer(function (callable $emit) use ($iterable, $delay) {
             foreach ($iterable as $value) {
+                if ($delay) {
+                    yield new Pause($delay);
+                }
+
                 yield $emit($value);
             }
         });
@@ -650,36 +653,6 @@ namespace Amp\Stream {
             }
 
             $emitter->resolve($values);
-        });
-
-        return $emitter->stream();
-    }
-
-    /**
-     * Returns a stream that emits a value every $interval milliseconds after (up to $count times). The value emitted
-     * is an integer of the number of times the stream emitted a value.
-     *
-     * @param int $interval Time interval between emitted values in milliseconds.
-     * @param int $count Number of values to emit. PHP_INT_MAX by default.
-     *
-     * @return \Amp\Stream
-     *
-     * @throws \Error If the number of times to emit is not a positive value.
-     */
-    function interval(int $interval, int $count = PHP_INT_MAX): Stream {
-        if (0 >= $count) {
-            throw new \Error("The number of times to emit must be a positive value");
-        }
-
-        $emitter = new Emitter;
-
-        Loop::repeat($interval, function ($watcher) use (&$i, $emitter, $count) {
-            $emitter->emit(++$i);
-
-            if ($i === $count) {
-                Loop::cancel($watcher);
-                $emitter->resolve();
-            }
         });
 
         return $emitter->stream();
